refactor: move referenceContent handling into repository
