+++ src/api/research-study/controllers/SearchResearchStudyController.ts
@@ -2,13 +2,12 @@ import { errors } from '@elastic/elasticsearch'
 import { Controller, Get, Header, Inject, Query, Res } from '@nestjs/common'
 import { ApiOkResponse, ApiOperation, ApiProduces, ApiTags } from '@nestjs/swagger'
 import { Response } from 'express'
-import { Bundle, BundleEntry, Group, Location, OperationOutcome, Organization } from 'fhir/r4'
+import { Bundle, OperationOutcome } from 'fhir/r4'
 
 import { ResearchStudyQueryParams } from './converter/ResearchStudyQueryParams'
 import { researchStudyQueryParamsToElasticsearchQuery } from './converter/researchStudyQueryParamsToElasticsearchQuery'
 import { ResearchStudyQueryModel } from './ResearchStudyQueryModel'
 import { ElasticsearchBodyType } from '../../../shared/elasticsearch/ElasticsearchBody'
-import { BundleEntryModel } from '../../../shared/models/backbone-elements/BundleEntryModel'
 import { OperationOutcomeModel } from '../../../shared/models/domain-resources/OperationOutcomeModel'
 import { ResearchStudyRepository } from '../application/contracts/ResearchStudyRepository'
 
@@ -31,20 +30,10 @@ export class SearchResearchStudyController {
 
       const elasticsearchBody: ElasticsearchBodyType = researchStudyQueryParamsToElasticsearchQuery(researchStudyQueryParams)
 
-      const withReferenceContents: boolean = researchStudyQueryParams.some(
-        (param) => param.name === '_include' && param.value === '*'
-      )
-
       const fhirResourceBundle: Bundle = await this.researchStudyRepository.search(
         elasticsearchBody,
-        researchStudyQueryParams,
-        withReferenceContents
+        researchStudyQueryParams
       )
-
-      if (withReferenceContents) {
-        const additionalFhirResourceBundle: BundleEntry[] = this.getAdditionalFhirResourceBundle(fhirResourceBundle)
-        fhirResourceBundle.entry.push(...additionalFhirResourceBundle)
-      }
 
       response.json(fhirResourceBundle)
     } catch (error) {
@@ -56,46 +45,5 @@ export class SearchResearchStudyController {
         throw error
       }
     }
-  }
-
-  private getAdditionalFhirResourceBundle(fhirResourceBundle: Bundle): BundleEntry[] {
-    const additionalFhirResourceBundleEntries = []
-
-    for (const bundleEntry of fhirResourceBundle.entry) {
-      const referenceContents: unknown = bundleEntry.resource['referenceContents']
-
-      const enrollmentGroup: Group = referenceContents['enrollmentGroup'] as Group
-      if (enrollmentGroup) {
-        const enrollmentGroupBundleEntry: BundleEntry = BundleEntryModel.create(enrollmentGroup, process.env['ECLAIRE_URL'])
-        additionalFhirResourceBundleEntries.push(enrollmentGroupBundleEntry)
-      }
-
-      const locations: Location[] = referenceContents['locations'] as Location[]
-      if (locations) {
-        const locationBundleEntries: BundleEntry[] = locations.map((location: Location) => {
-          return BundleEntryModel.create(location, process.env['ECLAIRE_URL'])
-        })
-
-        additionalFhirResourceBundleEntries.push(...locationBundleEntries)
-      }
-
-      const organizations: Organization[] = referenceContents['organizations'] as Organization[]
-      if (organizations) {
-        const organizationBundleEntries: BundleEntry[] = organizations.map((organization: Organization) => {
-          return BundleEntryModel.create(organization, process.env['ECLAIRE_URL'])
-        })
-
-        additionalFhirResourceBundleEntries.push(...organizationBundleEntries)
-      }
-
-      delete bundleEntry.resource['referenceContents']
-    }
-    // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access, @typescript-eslint/no-unsafe-return
-    const bundleIds = additionalFhirResourceBundleEntries.map((bundle) => bundle.resource.id)
-    const additionalFhirResourceBundleEntriesFiltered =
-      // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
-      additionalFhirResourceBundleEntries.filter((bundle, index) => !bundleIds.includes(bundle.resource.id, index + 1))
-
-    return [...additionalFhirResourceBundleEntriesFiltered as BundleEntry[]]
   }
 }
