+++ src/api/research-study/gateways/EsResearchStudyRepository.ts
@@ -1,8 +1,9 @@
 import { ConfigService } from '@nestjs/config'
-import { Bundle, BundleLink, ResearchStudy } from 'fhir/r4'
+import { Bundle, BundleEntry, BundleLink, Group, Location, Organization, ResearchStudy } from 'fhir/r4'
 
 import { ElasticsearchBodyType } from '../../../shared/elasticsearch/ElasticsearchBody'
 import { ElasticsearchService, SearchResponse } from '../../../shared/elasticsearch/ElasticsearchService'
+import { BundleEntryModel } from '../../../shared/models/backbone-elements/BundleEntryModel'
 import { BundleModel } from '../../../shared/models/resources/BundleModel'
 import { ResearchStudyRepository } from '../application/contracts/ResearchStudyRepository'
 import { ResearchStudyQueryParams } from '../controllers/converter/ResearchStudyQueryParams'
@@ -26,9 +27,12 @@ export class EsResearchStudyRepository implements ResearchStudyRepository {
 
   async search(
     elasticsearchBody: ElasticsearchBodyType,
-    queryParams: ResearchStudyQueryParams[],
-    withReferenceContents: boolean = false
+    queryParams: ResearchStudyQueryParams[]
   ): Promise<Bundle> {
+    const withReferenceContents: boolean = queryParams.some(
+      (param: ResearchStudyQueryParams) => param.name === '_include' && param.value === '*'
+    )
+
     const response: SearchResponse = await this.databaseService.search(elasticsearchBody, withReferenceContents)
 
     const links: BundleLink[] = []
@@ -38,12 +42,19 @@ export class EsResearchStudyRepository implements ResearchStudyRepository {
       this.buildSearchLinks(links, elasticsearchBody.from, response.total, queryParams)
     }
 
-    return BundleModel.create(
+    const fhirResourceBundle: Bundle = BundleModel.create(
       response.hits,
       links,
       response.total,
       this.configService.get('ECLAIRE_URL')
     )
+
+    if (withReferenceContents) {
+      const additionalFhirResourceBundle: BundleEntry[] = this.getAdditionalFhirResourceBundle(fhirResourceBundle)
+      fhirResourceBundle.entry.push(...additionalFhirResourceBundle)
+    }
+
+    return fhirResourceBundle
   }
 
   private buildSearchLinks(links: BundleLink[], offset: number, total: number, queryParams: ResearchStudyQueryParams[]) {
@@ -102,5 +113,46 @@ export class EsResearchStudyRepository implements ResearchStudyRepository {
     }
 
     return url.toString()
+  }
+
+  private getAdditionalFhirResourceBundle(fhirResourceBundle: Bundle): BundleEntry[] {
+    const additionalFhirResourceBundleEntries = []
+
+    for (const bundleEntry of fhirResourceBundle.entry) {
+      const referenceContents: unknown = bundleEntry.resource['referenceContents']
+
+      const enrollmentGroup: Group = referenceContents['enrollmentGroup'] as Group
+      if (enrollmentGroup) {
+        const enrollmentGroupBundleEntry: BundleEntry = BundleEntryModel.create(enrollmentGroup, this.configService.get('ECLAIRE_URL'))
+        additionalFhirResourceBundleEntries.push(enrollmentGroupBundleEntry)
+      }
+
+      const locations: Location[] = referenceContents['locations'] as Location[]
+      if (locations) {
+        const locationBundleEntries: BundleEntry[] = locations.map((location: Location) => {
+          return BundleEntryModel.create(location, this.configService.get('ECLAIRE_URL'))
+        })
+
+        additionalFhirResourceBundleEntries.push(...locationBundleEntries)
+      }
+
+      const organizations: Organization[] = referenceContents['organizations'] as Organization[]
+      if (organizations) {
+        const organizationBundleEntries: BundleEntry[] = organizations.map((organization: Organization) => {
+          return BundleEntryModel.create(organization, this.configService.get('ECLAIRE_URL'))
+        })
+
+        additionalFhirResourceBundleEntries.push(...organizationBundleEntries)
+      }
+
+      delete bundleEntry.resource['referenceContents']
+    }
+    // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access, @typescript-eslint/no-unsafe-return
+    const bundleIds = additionalFhirResourceBundleEntries.map((bundle) => bundle.resource.id)
+    const additionalFhirResourceBundleEntriesFiltered =
+      // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
+      additionalFhirResourceBundleEntries.filter((bundle, index) => !bundleIds.includes(bundle.resource.id, index + 1))
+
+    return [...additionalFhirResourceBundleEntriesFiltered as BundleEntry[]]
   }
 }
